:truck: api: move CertificationDetails read model to src
