+++ api/tests/tooling/domain-builder/factory/build-certification-details.js
@@ -1,5 +1,5 @@
-import { CertificationDetails } from '../../../../lib/domain/read-models/CertificationDetails.js';
 import { states } from '../../../../src/certification/session-management/domain/models/CertificationAssessment.js';
+import { CertificationDetails } from '../../../../src/certification/session-management/domain/read-models/CertificationDetails.js';
 
 const buildCertificationDetails = function ({
   id = 123,
